+++ packages/cln_common/example/cln_common_example.dart
@@ -1,0 +1,1 @@
+void main() {}
