+++ packages/cln_common/lib/cln_common.dart
@@ -1,0 +1,8 @@
+/// Support for doing something awesome.
+///
+/// More dartdocs go here.
+library cln_common;
+
+export 'src/cln_common_base.dart';
+
+// TODO: Export any libraries intended for clients of this package.
