+++ packages/cln_common/lib/src/json_rpc/error.dart
