+++ packages/cln_common/pubspec.yaml
@@ -1,19 +1,16 @@
-name: cln_plugin
+name: cln_common
 description: A starting point for Dart libraries or applications.
 version: 1.0.0
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.16.2 <3.0.0'
+  sdk: '>=2.17.3 <3.0.0'
 
 dependencies:
   json_annotation: ^4.5.0
-  cln_common:
-    path: ../cln_common
 
 dev_dependencies:
-  lints: ^1.0.0
+  lints: ^2.0.0
   test: ^1.16.0
   build_runner: ^2.0.0
   json_serializable: ^6.2.0
-  clightning_rpc: ^0.0.1
