+++ packages/rpc/pubspec.yaml
@@ -10,7 +10,9 @@ dev_dependencies:
   pedantic: ^1.10.0
   test: ^1.16.0
 
-dependencies: 
+dependencies:
   json_rpc_2: ^3.0.1
   jsonrpc2: ^3.0.1
   logger: ^1.0.0
+  cln_common:
+    path: ../cln_common
